Generation of content list in templates
The content list is now generated automatically in the templates. All
the files have to provide is the content. It is assumed a content list
will always be accompanied by a quick list.

Only the 404 page does not follow the template because it does not have
a quick list and its content list is somewhat different in appearance.
Therefore, even if most pages will have empty "content", the tag must
be kept in the templates.