Page names in French
At the moment, a redirection mechanism is used when a page that is
available with a French name is accessed using an English name. This is
probably not the best scheme, but it works. Accessing an English page
using a French name does not cause a redirection and causes a 404 error
instead.